refactor: simplify MemberAction by removing MemberForm and integrating its logic directly
